175 implement reboot support for agents

+++ lib/super_run_status.c
@@ -1,6 +1,6 @@
 /* -*- c -*- */
 
-/* Copyright (C) 2015-2022 Alexander Chernov <cher@ejudge.ru> */
+/* Copyright (C) 2015-2023 Alexander Chernov <cher@ejudge.ru> */
 
 /*
  * This program is free software; you can redistribute it and/or modify
@@ -99,7 +99,8 @@ super_run_status_save(
         long long *p_last_saved_time_ms,
         long long timeout_ms,
         unsigned char *p_stop_flag,
-        unsigned char *p_down_flag)
+        unsigned char *p_down_flag,
+        unsigned char *p_reboot_flag)
 {
     unsigned char in_path[PATH_MAX];
     unsigned char dir_path[PATH_MAX];
@@ -114,7 +115,8 @@ super_run_status_save(
     if (agent) {
         agent->ops->put_heartbeat(agent, file_name, psrs, sizeof(*psrs),
                                   p_last_saved_time_ms, p_stop_flag,
-                                  p_down_flag);
+                                  p_down_flag,
+                                  p_reboot_flag);
         return;
     }
 
@@ -161,6 +163,14 @@ super_run_status_save(
         snprintf(dir_path, sizeof(dir_path), "%s/dir/%s@D", heartbeat_dir, file_name);
         if (access(dir_path, F_OK) >= 0) {
             *p_down_flag = 1;
+            unlink(dir_path);
+        }
+    }
+    if (p_reboot_flag) {
+        *p_reboot_flag = 0;
+        snprintf(dir_path, sizeof(dir_path), "%s/dir/%s@R", heartbeat_dir, file_name);
+        if (access(dir_path, F_OK) >= 0) {
+            *p_reboot_flag = 1;
             unlink(dir_path);
         }
     }
